farm inventory section added

+++ database/migrations/2025_12_09_235312_create_expenses_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('expenses', function (Blueprint $table) {
             $table->id();
             $table->string('category');
-            $table->float('amount');
+            $table->decimal('amount', 20, 2);
             $table->string('payment_method');
             $table->date('date');
             $table->longText('description')->nullable();
